Move collection into mongo directory
