+++ mongo/collection.go
@@ -1,4 +1,4 @@
-package store
+package mongo
 
 import (
 	"context"
