+++ store/collection_operations_test.go
@@ -2,8 +2,8 @@ package store_test
 
 import (
 	"errors"
+	"github.com/ONSdigital/dp-files-api/mongo/mock"
 	"github.com/ONSdigital/dp-files-api/store"
-	"github.com/ONSdigital/dp-files-api/store/mock"
 	"github.com/ONSdigital/dp-kafka/v3/kafkatest"
 	mongodriver "github.com/ONSdigital/dp-mongodb/v3/mongodb"
 	"go.mongodb.org/mongo-driver/bson"
